Move mongocrypt.py out of async/sync
